pyproject.toml
[project]
name = "sales-forecasting-mlops"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "apache-airflow==3.0.1",
    "boto3>=1.42.3",
    "fastapi>=0.117.1",
    "holidays>=0.85",
    "lightgbm>=4.6.0",
    "loguru>=0.7.3",
    "matplotlib>=3.10.7",
    "mlflow==3.0.1",
    "numpy==1.26.4",
    "optuna>=4.6.0",
    "pandas>=2.3.3",
    "plotly>=6.5.0",
    "pyarrow>=4.0.0,<15.0.0",
    "pytest>=9.0.2",
    "scikit-learn>=1.7.2",
    "streamlit>=1.52.2",
    "uvicorn>=0.38.0",
    "xgboost-cpu>=3.1.2",
]
